GFF3 files should use species specific ids
As with BED files we should use a species specific id in GFF3 files.
This is consistent with what we show on the website as well as what we
now export in FASTA files. By using the same IDs it makes it easier to
compare data. Also, once we have location specific ids we should use
those instead.