Closer to working version of mapping import
This version does a better job of determining what to import. It can
detect the issues that can arise better but hasn't been tested with
actually inserting the new data yet. For now we will allow it simply
ignore the bad entries as they are a consequence of both partial imports
which are being cleaned up and the design of our database. This should
be cleaned up later.