Try to improve determination of coordinates
The basic idea is that there may be cross refs from more than one
database. In order to correctly update we have to select a single
database to use. SO to make this possible the user can now provide an
ordered choice of databases to update with. This seems to improve the
number of updates that are possible.