Simplify to not try and deduplicate
This is now a very simple exporter which will not attempt to dedupliate
in anyway. This is probably wrong overall but is useful for now. I will
come back later to better determine what should happen with each
database, plus where these files should be stored eventually. But for
now this is the kernel of the correct export.