Prevent obvious duplicates
This will prevent writing of rows which are completely duplicate. This
may not be smart enough for all the complex cases that can occur but is
a start. In addition, this assumes that the rows are ordered so that
duplicates appear next to each other. This is reasonable so long as the
query has the order by statement that uses all columns in the result.
This does impose an overhead on the database but is hopefully small
enough in practice to not matter.