Add more Ensembl genomes
This adds all the Ensembl genomes we don't currently have. This still
requires manual work to assign examples for all genomes as well as
determine the UCSC accession for all genomes.

I've also added a simple function to produce the location dictionary
based upon the string that Ensembl provides as a location. This makes
the cutting and pasting once I've chosen an example a lot simpler.