Add script to update known genomes
This adds a management command to query Ensembl's API to fetch the
descriptions of known genomes and store the required data in
portal.config.genomes. This allows the website to display genomes as
well as ftp export for those genomes (probably). The process does not
modify any already created genome, it will append new genome information
only. It also attempts to guess an example by selecting a random
accession that looks like it comes from a reasonable coordinate.

I have included a newly generated genomes.py.

Doing this required adding a chromosome attribute to the Accession
model.